Only purge WAN cache when edited on the network central wiki.

+++ includes/GloopTweaksHooks.php
@@ -211,26 +211,29 @@ class GloopTweaksHooks implements
 			CdnCacheUpdate::purgeGloop( [ "{$parsed['host']}{$parsed['path']}" ], 'prefix' );
 		}
 
-		if ( $wikiPage->getTitle()->getPrefixedDBkey() === 'MediaWiki:Robots.txt' ) {
-			// When [[MediaWiki:weirdgloop-contact-filter]] is edited, clear the 'robots' global cache key.
-			$cache = MediaWikiServices::getInstance()->getMainWANObjectCache();
+		$networkCentralDB = $this->config->get( 'GloopTweaksNetworkCentralDB' );
+		if ( $networkCentralDB && $this->config->get( MainConfigNames::DBname ) === $networkCentralDB ) {
+			if ( $wikiPage->getTitle()->getPrefixedDBkey() === 'MediaWiki:Robots.txt' ) {
+				// When [[MediaWiki:Robots.txt]] is edited, clear the 'robots' global cache key.
+				$cache = MediaWikiServices::getInstance()->getMainWANObjectCache();
 
-			$cache->delete(
-				$cache->makeGlobalKey(
-					'GloopTweaks',
-					'robots'
-				)
-			);
-		} elseif ( $wikiPage->getTitle()->getPrefixedDBkey() === 'MediaWiki:Weirdgloop-contact-filter' ) {
-			// When [[MediaWiki:weirdgloop-contact-filter]] is edited, clear the 'contact-filter-regexes' global cache key.
-			$cache = MediaWikiServices::getInstance()->getMainWANObjectCache();
+				$cache->delete(
+					$cache->makeGlobalKey(
+						'GloopTweaks',
+						'robots'
+					)
+				);
+			} elseif ( $wikiPage->getTitle()->getPrefixedDBkey() === 'MediaWiki:Weirdgloop-contact-filter' ) {
+				// When [[MediaWiki:Weirdgloop-contact-filter]] is edited, clear the 'contact-filter-regexes' global cache key.
+				$cache = MediaWikiServices::getInstance()->getMainWANObjectCache();
 
-			$cache->delete(
-				$cache->makeGlobalKey(
-					'GloopTweaks',
-					'contact-filter-regexes'
-				)
-			);
+				$cache->delete(
+					$cache->makeGlobalKey(
+						'GloopTweaks',
+						'contact-filter-regexes'
+					)
+				);
+			}
 		}
 	}
